pyproject.toml
[project]
name = "tg-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "black>=24.10.0",
    "pika>=1.3.2",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "python-socks[asyncio]>=2.5.3",
    "sanic-ext>=23.12.0",
    "sanic>=24.6.0",
    "telethon>=1.37.0",
]

[tool.uv]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
